code optimization and documentation
